refactor(samples): move extensions in a dedicated folder
