+++ ReduxSimple.Samples/Extensions/StringExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace ReduxSimple.Samples
+﻿namespace ReduxSimple.Samples.Extensions
 {
     public static class StringExtensions
     {
