+++ ReduxSimple.Samples/Pokedex/PokedexApiClient.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using ReduxSimple.Samples.Extensions;
 using System;
 using System.Net.Http;
 using System.Net.Http.Headers;
