+++ ReduxSimple.Samples/Pokedex/PokedexPage.xaml.cs
@@ -1,4 +1,5 @@
-﻿using SuccincT.Options;
+﻿using ReduxSimple.Samples.Extensions;
+using SuccincT.Options;
 using System;
 using System.Collections.Immutable;
 using System.Linq;
