+++ ReduxSimple.Samples/ReduxSimple.Samples.csproj
@@ -101,7 +101,7 @@
     </Compile>
     <Compile Include="Counter\CounterState.cs" />
     <Compile Include="Counter\CounterStore.cs" />
-    <Compile Include="IObservableExtensions.cs" />
+    <Compile Include="Extensions\IObservableExtensions.cs" />
     <Compile Include="MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
@@ -115,7 +115,7 @@
     <Compile Include="Pokedex\PokedexState.cs" />
     <Compile Include="Pokedex\PokedexStore.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="StringExtensions.cs" />
+    <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="TicTacToe\Actions.cs" />
     <Compile Include="TicTacToe\TicTacToePage.xaml.cs">
       <DependentUpon>TicTacToePage.xaml</DependentUpon>
@@ -131,7 +131,7 @@
     </Compile>
     <Compile Include="TodoList\TodoListState.cs" />
     <Compile Include="TodoList\TodoListStore.cs" />
-    <Compile Include="UIElementExtensions.cs" />
+    <Compile Include="Extensions\UIElementExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <AppxManifest Include="Package.appxmanifest">
